Divination/server/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
email-validator
tenacity
langchain
langchain-core
langchain-community
faiss-cpu
sentence-transformers
lunardate
google-generativeai
PyJWT
passlib[bcrypt]
